Add Makah (Qʷi·qʷidiččaq)
This adds the alphabet for the Makah language, a Wakashan language closely related to Nuu-chah-nulth and Diidiitidq.

I have been working with the community on keyboards (see: https://keyman.com/keyboards/nrc_makah) and this seems to be the cannonical set of characters used to type the language. I have also made a small font test application that tries the Makah alphabet in various fonts availabe in Google Docs: https://makah-fonts.vercel.app/

The biggest stuggle I face with the community is getting computers to render «x̌» and «ƛ̓» properly!